+++ packages/terra-collapsible-menu-view/src/CollapsibleMenuViewItemGroup.jsx
@@ -17,13 +17,24 @@ const propTypes = {
    */
   children: PropTypes.node.isRequired,
   /**
+   * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue) For best practices, ensure to use `isMultiSelect`
+   * in collapsibleMenuViewItemGroup element when creating multi select group items. This ensures collapsibleMenuItem's are rendered with role `checkbox` for face up group items and
+   * role `menuItemCheckBox` for collapsed group items.
+   */
+  isMultiSelect: PropTypes.bool,
+  /**
    * A list of keys of the CollapsibleMenuView.Items that should be selected.
    */
   selectedKeys: PropTypes.arrayOf(PropTypes.string),
+  /**
+   * String that labels the group for accessibility.
+   */
+  ariaLabel: PropTypes.string,
 };
 
 const defaultProps = {
   selectedKeys: [],
+  isMultiSelect: false,
 };
 
 const childContextTypes = {
@@ -37,14 +48,14 @@ const contextTypes = {
 class CollapsibleMenuViewItemGroup extends React.Component {
   constructor(props) {
     super(props);
-    this.handleOnChange = this.handleOnChange.bind(this);
+    this.handleMenuOnChange = this.handleMenuOnChange.bind(this);
   }
 
   getChildContext() {
     return { isCollapsibleGroupItem: true };
   }
 
-  handleOnChange(event, selectedIndex) {
+  handleMenuOnChange(event, selectedIndex) {
     if (this.props.onChange) {
       let selectedKey = selectedIndex;
       React.Children.forEach(this.props.children, (child, index) => {
@@ -61,25 +72,45 @@ class CollapsibleMenuViewItemGroup extends React.Component {
     const {
       children,
       onChange,
+      isMultiSelect,
       selectedKeys,
+      ariaLabel,
       ...customProps
     } = this.props;
 
     const { isCollapsibleMenuItem } = this.context;
 
-    if (isCollapsibleMenuItem && selectedKeys.length) {
+    if (isCollapsibleMenuItem) {
+      if (isMultiSelect) {
+        const cloneChildren = [];
+        React.Children.forEach(children, (item, index) => {
+          let clonedElement = item;
+          clonedElement = React.cloneElement(item, { onChange: (e) => this.handleMenuOnChange(e, index), isToggleable: true, isToggled: item.isToggled });
+          cloneChildren.push(clonedElement);
+        });
+        return (
+          <li role="none">
+            <div {...customProps} role="group">
+              {cloneChildren}
+            </div>
+          </li>
+        );
+      }
+      if (onChange) {
+        return (
+          <li role="none">
+            <Menu.ItemGroup {...customProps} onChange={this.handleMenuOnChange} aria-label={ariaLabel}>
+              {children}
+            </Menu.ItemGroup>
+          </li>
+        );
+      }
       return (
         <li role="none">
-          <Menu.ItemGroup {...customProps} onChange={this.handleOnChange}>
+          <div {...customProps} role="group">
             {children}
-          </Menu.ItemGroup>
+          </div>
         </li>
-      );
-    } if (isCollapsibleMenuItem) {
-      return (
-        <div {...customProps}>
-          { children }
-        </div>
       );
     }
 
@@ -89,7 +120,7 @@ class CollapsibleMenuViewItemGroup extends React.Component {
     ]);
 
     return (
-      <ButtonGroup {...customProps} onChange={onChange} className={buttonGroupClassNames} selectedKeys={selectedKeys}>
+      <ButtonGroup {...customProps} isMultiSelect={isMultiSelect} onChange={onChange} className={buttonGroupClassNames} selectedKeys={selectedKeys} aria-label={ariaLabel}>
         {children}
       </ButtonGroup>
     );