+++ packages/terra-framework-docs/src/terra-dev-site/doc/embedded-content-consumer/example/BasicConsumer.jsx
@@ -1,13 +1,17 @@
 import React from 'react';
 import EmbeddedContentConsumer from 'terra-embedded-content-consumer';
 import { Consumer } from 'xfc';
+import '../providers/EmbeddedContentConsumerCommon.module.scss';
 
 Consumer.init();
 
 const BasicConsumer = () => (
   <EmbeddedContentConsumer
     src="/terra-framework/#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/basic-provider"
-    options={{ iframeAttrs: { title: 'Basic content example' } }}
+    title="Basic content example"
+    options={{
+      resizeConfig: { scrolling: true },
+    }}
   />
 );