+++ packages/terra-framework-docs/src/terra-dev-site/doc/embedded-content-consumer/example/CustomEventConsumer.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import EmbeddedContentConsumer from 'terra-embedded-content-consumer';
 import { Consumer } from 'xfc';
+import '../providers/EmbeddedContentConsumerCommon.module.scss';
 
 Consumer.init();
 
@@ -15,7 +16,10 @@ const CustomEventConsumer = () => (
   <div id="CustomEvent">
     <EmbeddedContentConsumer
       src="/terra-framework/#/raw/provider/cerner-terra-framework-docs/embedded-content-consumer/providers/custom-event-provider"
-      options={{ iframeAttrs: { title: 'Custom Event Example' } }}
+      title="Custom Event Example"
+      options={{
+        resizeConfig: { scrolling: true },
+      }}
       eventHandlers={eventHandlers}
     />
   </div>