+++ packages/terra-framework-docs/src/terra-dev-site/doc/embedded-content-consumer/example/DataStatusConsumer.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import EmbeddedContentConsumer from 'terra-embedded-content-consumer';
 import { Consumer } from 'xfc';
+import '../providers/EmbeddedContentConsumerCommon.module.scss';
 
 Consumer.init();
 
@@ -23,7 +24,7 @@ const appendLifeCycleStatuses = (statuses) => {
   }
 };
 
-const options = { secret: 'OAuth Secret', iframeAttrs: { title: 'Embedded application lifecycle example', id: 'data-embedded-consumer-data-status' } };
+const options = { secret: 'OAuth Secret', iframeAttrs: { id: 'data-embedded-consumer-data-status' } };
 
 const onMount = () => {
   setTimeout(() => { appendLifeCycleStatuses(['Mounted']); }, 2000);
@@ -43,6 +44,7 @@ const DataStatusConsumer = () => (
     onMount={onMount}
     onLaunch={onLaunch}
     onAuthorize={onAuthorize}
+    title="Embedded application lifecycle example"
     options={options}
   />
 );