+++ packages/terra-tabs/src/common-tabs/_Tab.jsx
@@ -181,9 +181,6 @@ const Tab = ({
 
   if (isIconOnly) {
     attributes['aria-label'] = label;
-    if (isClosable && label) {
-      attributes['aria-label'] = `${label} . ${tabDeleteLabel}`;
-    }
   }
 
   function onCloseClick(event) {
@@ -260,6 +257,7 @@ const Tab = ({
             role="tab"
             className={variant === 'framework' ? paneClassNames : tabClassNames}
             title={label}
+            aria-label={isClosable ? `${label}. ${tabDeleteLabel}` : label}
             aria-describedby={responseId}
             tabIndex={isSelected ? 0 : -1}
             data-terra-tabs-show-focus-styles
@@ -274,8 +272,6 @@ const Tab = ({
             {isClosable && (
             <div
               className={cx('tabs-remove-button')}
-              type="button"
-              aria-label={`. ${tabDeleteLabel}`}
               onClick={onCloseClick}
             >
               <IconClose />
@@ -296,6 +292,7 @@ const Tab = ({
       aria-disabled={isDisabled}
       className={variant === 'framework' ? paneClassNames : tabClassNames}
       title={label}
+      aria-label={isClosable ? `${label}. ${tabDeleteLabel}` : label}
       tabIndex={isSelected ? 0 : -1}
       data-terra-tabs-show-focus-styles
     >
@@ -304,14 +301,14 @@ const Tab = ({
         {(!customDisplay && !isIconOnly) && <span className={variant === 'framework' ? cy('label') : cx('label')}>{label}</span>}
       </div>
       {isClosable && (
-      <div
-        className={cx('tabs-remove-button')}
-        type="button"
-        aria-label={`. ${tabDeleteLabel}`}
-        onClick={onCloseClick}
-      >
-        <IconClose />
-      </div>
+      <>
+        <div
+          className={cx('tabs-remove-button')}
+          onClick={onCloseClick}
+        >
+          <IconClose />
+        </div>
+      </>
       )}
     </div>
   );