+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed DateInput prop types for `excludeDates`, `includeDates`, `maxDate`, `minDate`.
+
 ## 4.90.0 - (September 19, 2023)
 
 * Fixed