+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## Unreleased
 
+* Added
+  * Added documentation for hyperlinks in `terra-collapsible-menu-view`.
+  * Added doc and test example for `collapsible-menu-view` multi-select group items.
+
+* Updated
+  * Updated examples for `terra-collapsible-menu-view` to use more meaningful controls and control labels.
+  * Updated examples and tests to add support for content scrolling within the iframe.
+  * Updated examples and tests for `terra-embedded-content-consumer` to show the preferred way to set the title.
+
+## 1.38.0 - (September 26, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.37.0 - (September 21, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.36.0 - (September 19, 2023)
 
 * Updated
@@ -10,6 +29,7 @@
 ## 1.35.0 - (September 14, 2023)
 
 * Added
+  * Added an example and test for `terra-embedded-content-consumer` to enable inline html content to be set in `srcdoc` frame attribute.
   * Added aria-label prop for terra-popup to announce the close instruction.
   * Added example for `terra-tabs` to test the focus traverse with interactive element within the tab content.
   * Added `terra-brand-footer` as a dependency.
@@ -57,7 +77,7 @@
 
 * Changed
   * Updated toggle examples for `terra-collapsible-menu-view` to include selectable toggles and toggles within a collapsed menu.
-  
+
 * Added
   * Updated row selection example for `terra-worklist-data-grid` to show ability to select the row selection column header.